[restructure] add setup and restructure project to turn it into an installable python packet
