+++ .gitignore
@@ -1,0 +1,4 @@
+psmqtt.egg-info
+*.pyc
+dist
+build
